change code lay out
